focus text edit and fix system ordering (#22)
# Objective

This PR makes immediate use of the console a bit more convenient by
focusing the text edit when it is opened. This allows you to ` and type
a command without clicking or tabbing to the text editor.

Additionally, this PR fixes a bug where the ordering of opening and
rendering the UI was indeterminate, causing the closing key press to
sometimes be sent to the text edit window. Now the ordering is fixed and
the console will always be opened or closed before it is rendered.

## Changelog

- The console text edit control will now be focused on console open.